refactor(utils/logout): move logout Fn to utils
